==> 01-postgis-store-parcels-blocks.ipynb <==
# Store parcels and blocks data in Postgres database

for a fast spatial-join of parcels to blocks.

First, create the database from terminal if it doesn't already exist:

```
createdb -U postgres parcels_blocks
psql -U postgres -d parcels_blocks -c "CREATE EXTENSION postgis;"
```

Save postgres username, password, host, port, and database name in keys.py. Then open the parcels and blocks shapefiles, then load each into tables in a PostGIS enabled database.

import geopandas as gpd
import psycopg2

# postgres info is stored in keys.py
from keys import pg_user, pg_pass, pg_host, pg_port, pg_db

crs = {'init':'epsg:4326'}

## Load and project the spatial data

%%time
# load blocks shapefile and project to 4326
blocks = gpd.read_file('data/tl_2010_06_tabblock10/')
blocks = blocks.to_crs(crs)
print(len(blocks))

%%time
# load parcels shapefile, drop any that lack geometry, and project to 4326
parcels = gpd.read_file('data/Parcels/')
parcels = parcels.dropna(subset=['geometry'])
parcels = parcels.to_crs(crs)
print(len(parcels))

# srid is the numeric spatial reference ID that PostGIS uses
assert crs == blocks.crs == parcels.crs
srid = crs['init'].strip('epsg:')
srid

## Clean up the parcels data

# convert integer columns to int
to_int = ['PARCEL_ID', 'DEVELOPMEN', 'COUNTY_ID', 'ZONE_ID', 'PROPORTION', 'TAX_EXEMPT', 'ID']
for col in to_int:
    parcels[col] = parcels[col].astype(int)

# drop unused columns and calculate centroids
parcels = parcels.drop(columns=['CENTROID', 'X', 'Y'])
parcels['centroid'] = parcels.centroid

## Insert blocks and parcels into PostGIS database

# connect to the postgres database
connection = psycopg2.connect(database=pg_db,
                              user=pg_user,
                              password=pg_pass,
                              host=pg_host,
                              port=pg_port)
cursor = connection.cursor()

# list all tables
cursor.execute("select relname from pg_class where relkind='r' and relname !~ '^(pg_|sql_)'")
cursor.fetchall()

### add blocks table and data

# drop blocks table if it already exists, then create blocks table
cursor.execute("DROP TABLE IF EXISTS blocks")
cursor.execute("""CREATE TABLE blocks (id SERIAL PRIMARY KEY,
                                       geoid VARCHAR NOT NULL)""")
cursor.execute("SELECT AddGeometryColumn ('blocks', 'geom', %s, 'MULTIPOLYGON', 2)", [srid])
cursor.execute("CREATE INDEX block_index ON blocks USING GIST(geom)")
connection.commit()

%%time
cursor.execute("DELETE FROM blocks")

# insert each block into the blocks table one at a time
for label, row in blocks.iterrows():
    geoid = row['GEOID10']
    geometry_wkt = row['geometry'].wkt
    
    query = """INSERT INTO blocks (geoid, geom) 
               VALUES (%s, ST_Multi(ST_GeomFromText(%s, %s)))"""
    data = (geoid, geometry_wkt, srid)
    cursor.execute(query, data)

connection.commit()

### add parcels table and data

# drop parcels table if it already exists, then create parcels table
cursor.execute("DROP TABLE IF EXISTS parcels")
cursor.execute("""CREATE TABLE parcels (id SERIAL PRIMARY KEY,
                                        parcel_id INTEGER NOT NULL)""")
cursor.execute("SELECT AddGeometryColumn ('parcels', 'geom', %s, 'POINT', 2)", [srid])
cursor.execute("CREATE INDEX parcel_index ON parcels USING GIST(geom)")
connection.commit()

%%time
cursor.execute("DELETE FROM parcels")

# insert each parcel into the parcels table one at a time
for label, row in parcels.iterrows():
    parcel_id = row['PARCEL_ID']
    geometry_wkt = row['centroid'].wkt
    
    query = """
            INSERT INTO parcels (parcel_id, geom)
            VALUES (%s, ST_GeomFromText(%s, %s))
            """
    data = (parcel_id, geometry_wkt, srid)
    cursor.execute(query, data)

connection.commit()

### optimize the database

%%time
# vacuum and analyze the database to optimize it after building indices and inserting rows
original_isolation_level = connection.isolation_level
connection.set_isolation_level(0)
cursor.execute("VACUUM ANALYZE")
connection.commit()
connection.set_isolation_level(original_isolation_level)

## Verify the data insertion

# verify the SRIDs
cursor.execute("""SELECT
                  Find_SRID('public', 'blocks', 'geom') as blocks_srid,
                  Find_SRID('public', 'parcels', 'geom') as parcels_srid""")
cursor.fetchall()

### verify the blocks

cursor.execute("SELECT count(*) AS exact_count FROM blocks")
rows = cursor.fetchall()
rows[0][0]

cursor.execute("SELECT geoid, ST_AsText(geom) FROM blocks LIMIT 3")
rows = cursor.fetchall()
gpd.GeoDataFrame(rows, columns=['GEOID', 'geometry'])

### verify the parcels

cursor.execute("SELECT count(*) AS exact_count FROM parcels")
rows = cursor.fetchall()
rows[0][0]

cursor.execute("""SELECT parcel_id, ST_AsText(geom)
                  FROM parcels LIMIT 3""")
rows = cursor.fetchall()
gpd.GeoDataFrame(rows, columns=['parcel_id', 'geometry'])

## All done

cursor.close()
connection.close()
