==> 02-postgis-join-parcels-blocks.ipynb <==
# Spatial join parcel IDs to census block IDs

import pandas as pd
import psycopg2
from keys import pg_user, pg_pass, pg_host, pg_port, pg_db

output_path = 'data/parcels_joined_blocks.csv'

# connect to the postgres database
connection = psycopg2.connect(database=pg_db,
                              user=pg_user,
                              password=pg_pass,
                              host=pg_host,
                              port=pg_port)
cursor = connection.cursor()

# verify that the SRIDs match
cursor.execute("""SELECT
                      Find_SRID('public', 'blocks', 'geom') as blocks_srid,
                      Find_SRID('public', 'parcels', 'geom') as parcels_srid""")
cursor.fetchall()

# joining how many rows to how many rows?
cursor.execute("SELECT count(*) AS row_count FROM blocks")
count_blocks_table = cursor.fetchall()[0][0]
cursor.execute("SELECT count(*) AS row_count FROM parcels")
count_parcels_table = cursor.fetchall()[0][0]
count_blocks_table, count_parcels_table

## Do the spatial join

%%time
# drop the table if it already exists and recreate it
# select distinct to keep only 1 parcel/block pair if parcel centroid lies on border of multiple blocks
query = """
        DROP TABLE IF EXISTS parcels_blocks;
        CREATE TABLE parcels_blocks AS
        SELECT DISTINCT ON (id)
            parcels.id AS id,
            parcels.parcel_id AS parcel_id,
            blocks.geoid AS block_geoid
        FROM parcels LEFT JOIN blocks
        ON ST_Intersects(parcels.geom, blocks.geom)
        """

cursor.execute(query)
connection.commit()

%%time
# vacuum and analyze the database to optimize it after creating new table
old_isolation_level = connection.isolation_level
connection.set_isolation_level(0)
cursor.execute("VACUUM ANALYZE")
connection.commit()
connection.set_isolation_level(old_isolation_level)

## Retrieve the joined data

# do all parcels have a block geoid?
cursor.execute("SELECT count(*) AS row_count, count(block_geoid) AS geoid_count FROM parcels_blocks")
print(count_parcels_table)
cursor.fetchall()

%%time
# select the results from the new table, ignoring any parcels with null block_geoid
cursor.execute("""SELECT parcel_id, block_geoid
                  FROM parcels_blocks
                  WHERE block_geoid IS NOT NULL""")
rows = cursor.fetchall()

# all done, close the database connection
cursor.close()
connection.close()

## Work with the results

# turn the selected rows into a dataframe
df = pd.DataFrame(rows, columns=['parcel_id', 'block_geoid'])
assert len(df) == count_parcels_table
len(df)

df.head()

df.to_csv(output_path, index=False, encoding='utf-8')
